+++ .env.example
@@ -1,0 +1,5 @@
+POSTGRES_USER=user
+POSTGRES_PASSWORD=password
+POSTGRES_SERVER=host_db
+POSTGRES_PORT=5432
+POSTGRES_DB=database_name
