+++ main.py
@@ -7,8 +7,6 @@ from routes import users, posts
 
 app = FastAPI()
 
-# models.Base.metadata.create_all(bind=engine)
-
 
 app.include_router(users.router)
 app.include_router(posts.router)
