+++ routes/posts.py
@@ -4,9 +4,7 @@ from fastapi import APIRouter, Depends, status, HTTPException
 
 from db import models, database, schemas
 from db.models import models
-from db.schemas import User
-from services import crud, oauth2
-
+from services import posts, oauth2, comments
 
 router = APIRouter(
     tags=['Blogs']
@@ -24,7 +22,7 @@ def create_post_for_user(
     Для создания поста
     """
     user = current_user
-    return crud.create_post(db=db, post=post, user=user)
+    return posts.create_post(db=db, post=post, user=user)
 
 
 @router.get("/posts/all")
@@ -32,7 +30,7 @@ def post_list(db: Session = Depends(get_db)):
     """
     Для вывода всех постов
     """
-    return crud.post_list(db=db)
+    return posts.post_list(db=db)
 
 
 @router.post("/posts/{post_id}/comment", response_model=schemas.CommentList)
@@ -44,7 +42,7 @@ def create_comment(comment: schemas.CommentBase,
     Для создании комментариев
     """
     user = current_user
-    return crud.create_comment(db=db, post_id=post_id, comment=comment, user=user)
+    return comments.create_comment(db=db, post_id=post_id, comment=comment, user=user)
 
 
 @router.get("/posts/{post_id}")
@@ -52,7 +50,7 @@ def post_detail(post_id: int, db: Session = Depends(get_db)):
     """
     Для получение инфо одного поста
     """
-    post = crud.get_posts(db=db, id=post_id)
+    post = posts.get_post(db=db, id=post_id)
     comments = db.query(models.Comment).options(
         joinedload(models.Comment.owner_comment).load_only("username", "email")
     ).filter(
