+++ services/comments.py
@@ -1,0 +1,14 @@
+from sqlalchemy.orm import Session
+
+from db import models, schemas
+
+
+def create_comment(db: Session, post_id: int, comment: schemas.CommentBase, user: schemas.User):
+    """
+    служебная функция для комментарии конктретного поста
+    """
+    db_comment = models.Comment(post_id=post_id, **comment.dict(), owner_comment=user)
+    db.add(db_comment)
+    db.commit()
+    db.refresh(db_comment)
+    return db_comment
