+++ services/posts.py
@@ -14,7 +14,7 @@ def create_post(db: Session, post: schemas.PostCreate, user: schemas.User):
     return db_post
 
 
-def get_posts(db: Session, id: int):
+def get_post(db: Session, id: int):
     """
     служебная функция отображение конктретного поста
     """
@@ -28,14 +28,3 @@ def post_list(db: Session):
     """
     return db.query(models.Post).options(joinedload(models.Post.owner).load_only(
         "username", "email")).all()
-
-
-def create_comment(db: Session, post_id: int, comment: schemas.CommentBase, user: schemas.User):
-    """
-    служебная функция для комментарии конктретного поста
-    """
-    db_comment = models.Comment(post_id=post_id, **comment.dict(), owner_comment=user)
-    db.add(db_comment)
-    db.commit()
-    db.refresh(db_comment)
-    return db_comment
